refactor(Library): separated out models, datasets, pipelines, experiments
