+++ .vscode/settings.json
@@ -7,7 +7,6 @@
     "python.linting.enabled": true,
     "python.formatting.provider": "black",
     "editor.formatOnSave": true,
-    "python.linting.mypyEnabled": true,
     "search.exclude": {
         "data/derived/**": true,
     },
