+++ blocks/transformations/__init__.py
@@ -2,3 +2,4 @@ from .lemmatizer import Lemmatizer
 from .sklearn import SKLearnTransformation
 from .text_statistics import TextStatisticTransformation
 from .tokenizer import SpacyTokenizer
+from .no_lemmatizer import NoLemmatizer
