+++ library/dataset/hatecheck.py
@@ -1,10 +1,9 @@
-from typing import Tuple
-
 import pandas as pd
 from datasets.arrow_dataset import Dataset
 
 from configs.constants import Const
-from type import PreprocessConfig, TestDataset, TrainDataset, DatasetSplit
+from type import PreprocessConfig, DatasetSplit
+from data.dataloader import DataLoader
 
 
 def transform_hatecheck_dataset(dataset: Dataset, config: PreprocessConfig) -> dict:
@@ -25,3 +24,13 @@ def transform_hatecheck_dataset(dataset: Dataset, config: PreprocessConfig) -> d
     )
 
     return {DatasetSplit.train.value: df_train, DatasetSplit.test.value: df_test}
+
+
+def get_hatecheck_dataloader() -> DataLoader:
+    return DataLoader("Paul/hatecheck", PreprocessConfig(
+        train_size=-1,
+        val_size=-1,
+        test_size=-1,
+        input_col="text",
+        label_col="label",
+    ), transform_hatecheck_dataset)
