+++ library/dataset/tweet_eval.py
@@ -1,15 +1,12 @@
-from typing import Tuple
-
 import pandas as pd
 from datasets.arrow_dataset import Dataset
 
 from configs.constants import Const
-from type import PreprocessConfig, TestDataset, TrainDataset, DatasetSplit
+from type import PreprocessConfig, DatasetSplit
+from data.dataloader import DataLoader
 
 
-def transform_dataset(
-    dataset: Dataset, config: PreprocessConfig
-) -> dict:
+def transform_dataset(dataset: Dataset, config: PreprocessConfig) -> dict:
 
     df_train = pd.DataFrame(dataset[DatasetSplit.train.value][: config.train_size])
     if DatasetSplit.val.value in dataset:
@@ -27,3 +24,18 @@ def transform_dataset(
     df_test = df_test.rename(columns=cols_to_rename)
 
     return {DatasetSplit.train.value: df_train, DatasetSplit.test.value: df_test}
+
+
+def get_tweet_eval_dataloader(name: str) -> DataLoader:
+    return DataLoader(
+        "tweet_eval",
+        PreprocessConfig(
+            train_size=-1,
+            val_size=-1,
+            test_size=-1,
+            input_col="text",
+            label_col="label",
+        ),
+        transform_dataset,
+        name,
+    )
