+++ library/models/sklearn_voting.py
@@ -1,0 +1,27 @@
+from sklearn.ensemble import GradientBoostingClassifier, VotingClassifier
+from sklearn.feature_extraction.text import TfidfVectorizer
+from sklearn.linear_model import LogisticRegression
+from sklearn.naive_bayes import MultinomialNB
+from type import SKLearnConfig
+from configs.constants import Const
+
+sklearn_config = SKLearnConfig(
+    frozen=False,
+    save=True,
+    preferred_load_origin=None,
+    classifier=VotingClassifier(
+        estimators=[
+            ("nb", MultinomialNB()),
+            ("lg", LogisticRegression(random_state=Const.seed)),
+            (
+                "gb",
+                GradientBoostingClassifier(
+                    n_estimators=100, max_depth=20, random_state=Const.seed
+                ),
+            ),
+        ],
+        voting="soft",
+    ),
+    one_vs_rest=False,
+    save_remote=False,
+)
